Add files via upload
This zip file contains all the files required for this project. As we are connected to the local host, we can access them, and I'll attach the database files also. If you want to check out, you have to download the DB (database) files, and you can also access the project, and this project is done in PHP.